--- tools/cli/collection.json
@@ -1,10 +1,0 @@
-{
-  "$schema": "../../node_modules/@angular-devkit/schematics/collection-schema.json",
-  "schematics": {
-    "ng-add": {
-      "description": "Add Daffodil ecommerce capabilities to an existing Angular project",
-      "factory": "./ng-add/index#ngAdd",
-      "schema": "./ng-add/schema.json"
-    }
-  }
-}
