--- tools/cli/scripts/copy-package-json.js
@@ -1,7 +1,0 @@
-const { readFileSync, writeFileSync } = require('fs');
-
-const pkg = JSON.parse(readFileSync(`${__dirname}/../package.json`, 'utf-8'));
-delete pkg.devDependencies;
-delete pkg.nx;
-delete pkg.scripts;
-writeFileSync(`${__dirname}/../../../dist/cli/package.json`, JSON.stringify(pkg, null, 2) + '\n');
