--- tools/cli/tsconfig.json
@@ -1,36 +1,0 @@
-{
-  "extends": "../tsconfig.tools.json",
-  "compilerOptions": {
-    "outDir": "../../dist/cli",
-    "rootDir": ".",
-    "module": "node16",
-		"moduleResolution": "node16",
-    "target": "es2022",
-    "lib": ["es2022"],
-    "declaration": true,
-    "experimentalDecorators": true,
-    "importHelpers": false,
-    "skipLibCheck": true,
-    "types": ["node"],
-		"resolveJsonModule": true,
-		"paths": {
-      "@daffodil/*": [
-        "dist/*"
-      ],
-      "@daffodil/cli": [
-        "libs/cli/src"
-      ],
-      "@daffodil/cli/versioning": [
-        "libs/cli/versioning"
-      ],
-    }
-  },
-  "include": [
-    "**/*.ts",
-  ],
-  "exclude": [
-    "**/*.spec.ts",
-    "**/files/**/*.ts",
-    "node_modules"
-  ],
-}
