--- tools/cli/tsconfig.spec.json
@@ -1,20 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "../../out-tsc/spec",
-		"module": "commonjs",
-		"moduleResolution": "node",
-    "types": [
-      "jasmine",
-      "node"
-    ],
-    "esModuleInterop": true,
-    "allowSyntheticDefaultImports": true,
-    "skipLibCheck": true,
-    "resolveJsonModule": true
-  },
-  "include": [
-    "**/*.spec.ts",
-    "**/*.d.ts"
-  ]
-}
