+++ tools/schematics/scripts/build.sh
@@ -7,6 +7,7 @@ tsc -p tsconfig.json
 
 cp collection.json "$DIST/"
 cp ng-add/schema.json "$DIST/ng-add/"
+cp versioning/daff.schema.json "$DIST/"
 cp -r ng-add/files "$DIST/ng-add/"
 cp README.md "$DIST/"
