+++ tools/schematics/scripts/collect-platform-versions.mts
