+++ tools/schematics/versioning/magento/find-supported.ts
