+++ tools/schematics/versioning/magento/parse.ts
