+++ tools/schematics/versioning/magento/stringify.spec.ts
