+++ tools/schematics/versioning/magento/stringify.ts
