+++ tools/schematics/versioning/semantic/stringify.spec.ts
