+++ tools/schematics/versioning/semantic/stringify.ts
