+++ tools/schematics/versioning/sync-projects.spec.ts
