+++ tools/schematics/versioning/sync-projects.ts
