+++ tools/schematics/versioning/validate-platform.ts
